--- fbx/builder.go
@@ -1,3 +1,0 @@
-package fbx
-
-const bufSize = 2 << 10
